ofproto-dpif-xlate: Add a recursion limit to tunnel address lookup.
In deployments with multiple tunnels, it can be possible to enter a an
infinite loop where traffic generates an arp/nd lookup which is
forwarded to a different tunnel, generating a new arp/nd packet that is
send back to the first tunnel. Now the depth counter is incremented for
new arp/nd packets, just as happens in xlate_table_action().

Signed-off-by: Mike Pattrick <mkp@redhat.com>