ovs-lib: Recreate db when schema is corrupted.
Previously ovs-lib would assume a database is valid if the file exists,
however, it is possible for the database file to exist but not be valid
for ovsdb to open it.

Now existence checks are augmented with schema checksum validation.
Databases with an invalid schema are removed.

Reported-at: https://issues.redhat.com/browse/FDP-689
Reported-by: Ihar Hrachyshka
Signed-off-by: Mike Pattrick <mkp@redhat.com>