ipf: Maintain conntrack zone.
Currently ipf will inject completed fragments into the first available
batch. In almost all cases, this is the batch which contained the last
fragment of the packet. However, in cases where the batch is full the
packets are added to whatever random subsequent batch arrives to
conntrack. This could result in packets being sent to an inappropriate
interface.

This patch verifies the zone matches, and that the batch contains a
packet of the same in_port as the completed fragments.

https://issues.redhat.com/browse/FDP-1052
Fixes: 4ea96698f667 ("Userspace datapath: Add fragmentation handling.")
Signed-off-by: Mike Pattrick <mkp@redhat.com>

Signed-off-by: Mike Pattrick <mkp@redhat.com>