dpif-netdev: Add PMTU support for tunnels.
Add support in the userspace datapath for PATH MTU on tunnel interfaces.

This feature allows users to configure an MTU on tunnel ports. If set,
when the userspace datapath attempts to encapsulate a packet that
exceeds the tunnels MTU, OVS will generate and send an ICMP
Fragmentation Needed or Packet Too Big message back to the source host.

If an MTU is not set on the tunnel interface, there is no change in
behaviour.

Reported-at: https://issues.redhat.com/browse/FDP-256
Signed-off-by: Mike Pattrick <mkp@redhat.com>